Combine ras_hdf.py and ras_stac.py
